llm-finetune-sft-dpo/requirements.txt

# Core training stack
transformers>=4.43
accelerate>=0.33
datasets>=2.20
peft>=0.12
trl>=0.9.4

# Serving
fastapi>=0.111
uvicorn>=0.30

# Tracking
mlflow>=2.14

# Optional eval
lm-eval==0.4.3

# Utils
pydantic>=2.7
PyYAML>=6.0
rich>=13.7
